refactor: remove unused endpoints and add more tests

+++ app/Actions/CreateAvatarAction.php
@@ -16,7 +16,7 @@ final readonly class CreateAvatarAction
     public function handle(User $user, UploadedFile $avatar): User
     {
         return DB::transaction(function () use ($user, $avatar): User {
-            if ($user->profile?->avatar) {
+            if ($user->profile?->avatar && $user->profile->avatar !== Constants::DEFAULT_PROFILE_PICTURE_PATH) {
                 Storage::disk('public')->delete($user->profile->avatar);
             }
